@@ -10,63 +10,62 @@ using VKirienko.Web.Services;
 using VKirienko.Web.Settings;
 using VKirienko.Web.SignalR;
 
-namespace VKirienko.Web
+namespace VKirienko.Web;
+
+public class Startup(IConfiguration configuration)
 {
-    public class Startup(IConfiguration configuration)
+    public IConfiguration Configuration { get; } = configuration;
+
+    // This method gets called by the runtime. Use this method to add services to the container.
+    public void ConfigureServices(IServiceCollection services)
     {
-        public IConfiguration Configuration { get; } = configuration;
+        var settings = Configuration.GetSection("Settings").Get<ApplicationSettings>();
 
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
-        {
-            var settings = Configuration.GetSection("Settings").Get<ApplicationSettings>();
+        services.AddSingleton(settings);
 
-            services.AddSingleton(settings);
+        services.AddApplicationInsightsTelemetry();
+        services.AddMvc();
+        services.AddControllersWithViews();
+        services.AddRazorPages();
+        services.AddSignalR();
 
-            services.AddApplicationInsightsTelemetry();
-            services.AddMvc();
-            services.AddControllersWithViews();
-            services.AddRazorPages();
-            services.AddSignalR();
-
-            services.AddAutoMapper(typeof(Startup));
+        services.AddAutoMapper(cfg => { }, typeof(Startup));
 
 #if DEBUG
-            services.AddSingleton<TimerManager>();
+        services.AddSingleton<TimerManager>();
 #endif
-            services.AddScoped<ITelemetryService, TelemetryService>();
+        services.AddScoped<ITelemetryService, TelemetryService>();
 
-            services.AddDbContext<IoTContext>(options =>
-                options.UseSqlite(Configuration.GetConnectionString("IoTDatabase")));
-        }
+        services.AddDbContext<IoTContext>(options =>
+            options.UseSqlite(Configuration.GetConnectionString("IoTDatabase")));
+    }
 
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    {
+        if (env.IsDevelopment())
         {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-            else
-            {
-                app.UseExceptionHandler("/Error");
-            }
-
-            app.UseForwardedHeaders(new ForwardedHeadersOptions
-            {
-                ForwardedHeaders = ForwardedHeaders.XForwardedFor | ForwardedHeaders.XForwardedProto
-            });
-
-            app.UseHttpsRedirection();
-            app.UseStaticFiles();
-            app.UseRouting();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllerRoute("default", "{controller}/{action=Index}/{id?}");
-                endpoints.MapHub<TelemetryHub>("/telemetry");
-                endpoints.MapFallbackToFile("index.html");
-            });
+            app.UseDeveloperExceptionPage();
         }
+        else
+        {
+            app.UseExceptionHandler("/Error");
+        }
+
+        app.UseForwardedHeaders(new ForwardedHeadersOptions
+        {
+            ForwardedHeaders = ForwardedHeaders.XForwardedFor | ForwardedHeaders.XForwardedProto
+        });
+
+        app.UseHttpsRedirection();
+        app.UseStaticFiles();
+        app.UseRouting();
+
+        app.UseEndpoints(endpoints =>
+        {
+            endpoints.MapControllerRoute("default", "{controller}/{action=Index}/{id?}");
+            endpoints.MapHub<TelemetryHub>("/telemetry");
+            endpoints.MapFallbackToFile("index.html");
+        });
     }
 }
